+++ server/service/interfaces/group_interface.go
