+++ server/service/interfaces/message_interface.go
